Move CanFrame to itself own module
